+++ app/javascript/stylesheets/drivers_order_deliver.scss
@@ -1,0 +1,61 @@
+.driver_profiles.index {
+  background-color: #18191B;
+  .map-container {
+    @apply flex flex-col pt-0;
+    height: calc(100vh - 70px);
+    .destination {
+      @apply absolute z-10 bg-white;
+    }
+    #map {
+      @apply flex-grow w-full;
+      margin-bottom: -25px;
+    }
+    .steps {
+      @apply w-full z-10;
+      background-color: rgb(9, 95, 35);
+      color: rgb(255, 255, 255);
+      padding: 10px;
+    }  
+    .info {
+      @apply flex-none rounded-t bg-white relative;
+      .status {
+        @apply text-center mt-5;
+        p {
+          @apply font-medium text-2xl;
+        }
+        .online-btn {
+          @apply bg-red-500 absolute right-2 top-2;
+        }
+      }
+      .order {
+        @apply pt-5 px-2 border-t-2;
+        .distance-matrix {
+          @apply flex font-bold text-2xl justify-center;
+        }
+        .title {
+          @apply flex;
+          .store-name {
+            @apply font-medium text-2xl mr-5;
+          }
+          .phone {
+            a {
+              @apply text-red-500 text-2xl;
+            }
+          }
+        }
+        .store-address, .order-number, .orderer, .order-address {
+          @apply font-light text-sm text-gray-500;
+        }
+        .text {
+
+        }
+        .btn-list {
+          @apply flex;
+        }
+      }
+    }
+    .hidden {
+      @apply invisible;
+    }
+  }
+}
