+++ app/views/driver_profiles/order_deliver.json.jbuilder
@@ -1,0 +1,4 @@
+json.set! :order do
+  json.extract! @order, :username, :tel, :address, :num
+  json.extract! @store, :store_name, :store_address, :store_phone
+end
