+++ config/routes.rb
@@ -42,6 +42,7 @@ Rails.application.routes.draw do
 
   resource :driver_profiles, path: '/drivers', only: [:new, :create, :edit, :update] do
     get :index
+    get :order_deliver
     post :online
   end
